+++ ios/Classes/VideoThumbKitPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface VideoThumbKitPlugin : NSObject<FlutterPlugin>
-@end