+++ windows/video_thumb_kit_plugin.h
@@ -1,14 +1,15 @@
 #ifndef FLUTTER_PLUGIN_VIDEO_THUMB_KIT_PLUGIN_H_
 #define FLUTTER_PLUGIN_VIDEO_THUMB_KIT_PLUGIN_H_
 
-#include <flutter/method_channel.h>
 #include <flutter/plugin_registrar_windows.h>
 
 #include <memory>
 
+#include "messages.g.h"
+
 namespace video_thumb_kit {
 
-class VideoThumbKitPlugin : public flutter::Plugin {
+class VideoThumbKitPlugin : public flutter::Plugin, public VideoThumbKitHostApi {
  public:
   static void RegisterWithRegistrar(flutter::PluginRegistrarWindows* registrar);
 
@@ -18,10 +19,13 @@ class VideoThumbKitPlugin : public flutter::Plugin {
   VideoThumbKitPlugin(const VideoThumbKitPlugin&) = delete;
   VideoThumbKitPlugin& operator=(const VideoThumbKitPlugin&) = delete;
 
- private:
-  void HandleMethodCall(
-      const flutter::MethodCall<flutter::EncodableValue>& method_call,
-      std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result);
+  // VideoThumbKitHostApi
+  void GenerateThumbnailFile(
+      const ThumbnailRequest& request,
+      std::function<void(ErrorOr<std::optional<std::string>> reply)> result) override;
+  void GenerateThumbnailData(
+      const ThumbnailRequest& request,
+      std::function<void(ErrorOr<std::optional<std::vector<uint8_t>>> reply)> result) override;
 };
 
 }  // namespace video_thumb_kit